Completion Group Management
Users can now create / edit groups , set permissions and members to
their tables.

+++ login/views.py
@@ -3,10 +3,7 @@ from django.shortcuts import get_object_or_404, render, redirect,\
 from django.views.generic import View
 from django.views.generic.edit import UpdateView
 from .models import myuser as OepUser
-from rest_framework.authtoken.models import Token
-from django.template.context import RequestContext
-from .forms import GroupPermForm, ListGroups, AllPermForm
-from django.contrib.admin.helpers import Fieldset
+from .forms import AllPermForm, GroupUserForm
 from django.contrib.auth.models import Group
 
  
@@ -31,22 +28,45 @@ class ProfileView(View):
 
 class GroupManagement(View):
     def get(self, request, user_id):
+        """
+        Load and list the available groups by groupadmin. 
+        :param request: A HTTP-request object sent by the Django framework.
+        :param user_id: An user id
+        :return: Profile renderer   
+        """
         user = get_object_or_404(OepUser, pk=user_id)
         groups = user.groupadmin.all()       
         return render(request, "login/admin_group.html", {'user': user, 'groupresult': groups})
     
 class GroupEdit(View):
     def get(self, request, user_id, group_id=""):
+        """
+        Load the chosen action(create or edit) for a group. 
+        :param request: A HTTP-request object sent by the Django framework.
+        :param user_id: An user id
+        :param user_id: An group id
+        :return: Profile renderer   
+        """
         user = get_object_or_404(OepUser, pk=user_id)
+        members = GroupUserForm(group_id=group_id)
         if group_id != "":
             group = get_object_or_404(Group, pk=group_id)
             form = AllPermForm(user = user, group = group)
             form.label = group.name
-            return render(request, "login/change_form.html", {'user': user, 'group_id': group_id, 'form': form})
+            return render(request, "login/change_form.html", {'user': user, 'group_id': group_id, 'form': form, 'members':members})    
         form = AllPermForm(user = user, group = "")
-        return render(request, "login/change_form.html", {'user': user, 'form': form})
+        return render(request, "login/change_form.html", {'user': user, 'form': form, 'members':members})
         
     def post(self, request, user_id, group_id=""):
+        """
+        Performs selected action(save or delete) for a group. If a groupname already exists, then a error 
+        will be output. 
+        The selected users become members of this group. The groupadmin is already set.
+        :param request: A HTTP-request object sent by the Django framework.
+        :param user_id: An user id
+        :param user_id: An group id
+        :return: Profile renderer   
+        """
         user = get_object_or_404(OepUser, pk=user_id)
         if group_id != "":
             group = get_object_or_404(Group, pk=group_id)
@@ -54,34 +74,47 @@ class GroupEdit(View):
         else:
             form = AllPermForm(request.POST, user = user, group = "")
         form.label = request.POST["group_name"]
-        print(form.is_valid())
-        if form.is_valid():
+        members = GroupUserForm(request.POST, group_id=group_id)
+
+        if form.is_valid() and members.is_valid():
             groupperms = form.cleaned_data.get('allperms')
-            print(groupperms)
+            groupuser = members.cleaned_data.get('groupmembers')
+            
             if "Save" in request.POST["submit"]:
-                print('save')
                 if group_id != "":
                     if groupperms == None:
                         group.permissions.clear()
                     else:
-                        group.permissions.set(groupperms) 
+                        group.permissions.set(groupperms)
+                    _set_group_members(user=user, group=group, groupuser=groupuser)                 
                     Group.objects.filter(id = group_id).update(name = form.label)
                 else:
+                    if Group.objects.filter(name = form.label).exists():
+                        error = "Groupname already exists! Chose another name. "
+                        return render(request, "login/change_form.html", {'user': user, 'form': form, 'error': error})
+                    
                     group = Group.objects.create(name = form.label)
                     if groupperms != None:
                         group.permissions.set(groupperms)
                     user.groupadmin.add(group)
-                    user.groups.add(group)
-                        
-        
+                    _set_group_members(user=user, group=group, groupuser=groupuser)
+                   
             elif "Delete" in request.POST["submit"]:
                 group.permissions.clear()
                 user.groupadmin.remove(group)
                 Group.objects.filter(id = group_id).delete()
-                print('delete')
+
         groups = user.groupadmin.all()  
         return render(request, "login/admin_group.html", {'user': user, 'groupresult': groups})
-    
+
+def _set_group_members(user, group, groupuser):
+    """
+    Help function for EditGroups. This set the selected users to the group.
+    """ 
+    group.user_set.clear()
+    group.user_set.set(groupuser)
+    user.groups.add(group)
+
 class ProfileUpdateView(UpdateView):
     """
     Autogenerate a update form for users.
